+++ CHANGELOG.md
@@ -12,12 +12,16 @@ All notable changes to this project will be documented in this file. This projec
 - **Settings**: remove get command [#135][]
 - **Changes**: rename --remote to --upstream [#140][]
 
+### Fixes
+- **State**: multiple --no-show-* options work together [#151][]
+
 [#54]: https://github.com/Brickster/git-commands/issues/54
 [#118]: https://github.com/Brickster/git-commands/issues/118
 [#134]: https://github.com/Brickster/git-commands/issues/134
 [#135]: https://github.com/Brickster/git-commands/issues/135
 [#140]: https://github.com/Brickster/git-commands/issues/140
 [#143]: https://github.com/Brickster/git-commands/issues/143
+[#151]: https://github.com/Brickster/git-commands/issues/151
 
 ## [v0.7.0][] - 2019-08-26
 ### Features